refactor(04-react-prueba-tecnica): divide logic
