+++ projects/04-react-prueba-tecnica/src/components/Otro.jsx
@@ -1,0 +1,10 @@
+import { useCatImage } from '../hooks/useCatImage.js'
+
+export function Otro(){
+    const { imageUrl } = useCatImage({ fact: 'cat' })
+    return (
+        <>
+            {imageUrl && <img src={imageUrl} />}
+        </>
+    )
+}
